ai-service/requirements.txt
# AI Analysis Layer - Dependencies
# Updated for Python 3.14+ compatibility

# Web Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0

# Settings & Validation
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Async Support
anyio>=4.2.0

# OpenAI Integration
openai>=1.0.0

# Vector DB client (S10 / F12 RAG Mentor Chat — ADR-036)
qdrant-client>=1.13.0,<2.0.0

# Environment variables
python-dotenv>=1.0.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.23.0
httpx>=0.26.0

# Security analyzer (for Python analysis)
bandit>=1.7.0

# File upload support
python-multipart>=0.0.6

# PDF generation (kept for future use)
reportlab>=4.0.0
